backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.1
pydantic==2.5.0
python-multipart==0.0.6
sqlalchemy==2.0.23
firebase-admin==6.2.0
stripe==7.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
beautifulsoup4==4.12.2
requests==2.31.0
pypdf2==3.0.1
python-docx==1.1.0

